+++ cdk-postgresql/lib/lambda.types.ts
@@ -1,0 +1,21 @@
+import { CloudFormationCustomResourceCreateEvent } from "aws-lambda";
+
+export interface Connection {
+  Host: string;
+  Port: number;
+  Username: string;
+  PasswordArn: string;
+  PasswordField?: string;
+  Database: string;
+  SSLMode: "require" | "disable";
+}
+
+export interface CreateDatabaseEvent
+  extends CloudFormationCustomResourceCreateEvent {
+  ResourceProperties: {
+    ServiceToken: string;
+    Connection: Connection;
+    Name: string;
+    Owner: string;
+  };
+}
