+++ cdk-postgresql/test/database.test.ts
@@ -1,14 +1,16 @@
-import { CloudFormationCustomResourceEvent } from "aws-lambda";
 import { handler } from "../lib/database.handler";
 import { GenericContainer, StartedTestContainer } from "testcontainers";
 import ms from "ms";
+import { CreateDatabaseEvent } from "../lib/lambda.types";
+
+const DB_PORT = 5432;
 
 describe("database", () => {
   let container: StartedTestContainer;
 
   beforeAll(async () => {
     container = await new GenericContainer("postgres")
-      .withExposedPorts(5432)
+      .withExposedPorts(DB_PORT)
       .withEnv("POSTGRES_PASSWORD", "postgres")
       .start();
   }, ms("2m"));
@@ -19,7 +21,7 @@ describe("database", () => {
 
   test("something", async () => {
     console.log({ handler });
-    const event: CloudFormationCustomResourceEvent = {
+    const event: CreateDatabaseEvent = {
       RequestType: "Create",
       ServiceToken: "",
       ResponseURL: "",
@@ -29,7 +31,16 @@ describe("database", () => {
       ResourceType: "",
       ResourceProperties: {
         ServiceToken: "",
-        Connection: { Host: container.getHost() },
+        Connection: {
+          Host: container.getHost(),
+          Port: container.getMappedPort(DB_PORT),
+          Username: "postgres",
+          Database: "postgres",
+          PasswordArn: "asdf",
+          SSLMode: "disable",
+        },
+        Name: "mydb",
+        Owner: "postgres",
       },
     };
     await handler(event);
